last up s1

+++ src/main/java/com/ssa/controller/NotifRestController.java
@@ -25,7 +25,7 @@ public class NotifRestController {
     
     @GetMapping("/notifNumber")
     public int getNotifNb(){
-        return notificationRepository.findByReportedIsFalse().size();
+        return notificationRepository.findByReportedIsFalseAndDisabledFalse().size();
     }
     
     @GetMapping("/projetNumber")
@@ -35,13 +35,12 @@ public class NotifRestController {
     
     @GetMapping("/notifList")
     public List<Notification> getNotifls(){
-        
-        return notificationRepository.findByReportedIsFalse();
+        return notificationRepository.findByReportedIsFalseAndDisabledFalse();
     }
     
     @GetMapping("/notif_reported_list")
     public List<Notification> getNotiflsrrr(){
-        return notificationRepository.findByReportedIsTrue();
+        return notificationRepository.findByReportedIsTrueAndDisabledFalse();
     }
     
     @GetMapping("/projetList")
@@ -49,14 +48,15 @@ public class NotifRestController {
         List<Projet> projetList= projetService.findAll();
         projetList.stream().forEach(projet -> {
             projet.setSuiviChantier(null);
-            projet.setDescProjet(null); });
+            projet.setDescProjet(null);
+        });
         return projetList;
     }
     @GetMapping("/{ref}/notif")
     public int getprhh(@PathVariable("ref") String ref){
         if(notificationRepository.findByProjetid(ref) == null){
             Projet projet= projetService.findById(ref);
-            Notification notification= new Notification(0, ref, projet.getFicheProjet().getPrenom()+" "+projet.getFicheProjet().getNom(), new Date(), new Date(), false);
+            Notification notification= new Notification(0, ref, projet.getFicheProjet().getPrenom()+" "+projet.getFicheProjet().getNom(), new Date(), new Date(), false, false);
             notificationService.save(notification);
         }
         return 1;
@@ -76,7 +76,7 @@ public class NotifRestController {
     @GetMapping("/{ref}/notif/delete")
     public int getprhdhg(@PathVariable("ref") String ref){
         Notification notification= notificationRepository.findByProjetid(ref);
-        notificationService.deleteById(notification.getId());
+        if(notification != null){ notificationService.deleteById(notification.getId()); }
         return 1;
     }
